reorganized dispatcher a bit
